Renames UriReference to UnixFilePath and changes its regex accordingly. These were never references to URIs.
